+++ tests/bench/b_strictfold.hs
@@ -1,0 +1,4 @@
+import Data.List (foldl')
+
+main :: IO ()
+main = print (foldl' (+) 0 [1 .. 2000000 :: Integer])